METAR incluir mapa con bias

+++ METAR/get-info-METARandECMWF.sh
@@ -1,71 +1,83 @@
 #!/bin/bash
 ####################################################################################
-### Descarga METARs info y las predicciones del ECMWF en la localidad de los METARs 
-### Compara temperaturas
-### Pinta series con R
+### Validación de predicciones meteored con METARs
+### 1. Descarga METARs info y las predicciones del ECMWF en la localidad de los METARs
+### 2. Compara temperaturas
+### 3. Pinta series con R y mapa de sesgo
 ### RLP 11-2019
 ###################################################################################
 # Nombre del script.
-scriptName=$(basename $0)
-PATHsoft1=/usr/bin
-PATHsoft2=/usr/bin
+scriptName=$(basename "$0")
+
 # Formato de fecha para el fichero .log.
 function datePID {
 	  echo "$(date -u +%Y/%m/%d\ %H:%M:%S) UTC [$$]"
   }
 
 # Comprobación de que existe software para optimizar.
-#FIXME solo dejo el nombre del software sin el path (diferencia entre command y which?)
-Rscript="/usr/bin/Rscript"
-pdfjoin="/usr/bin/pdfjoin"
-command -v ${Rscript} > /dev/null 2>&1 || { echo "$(datePID): ${Rscript} no está instalado." && exit 1; }
-command -v ${pdfjoin} > /dev/null 2>&1 || { echo "$(datePID): ${pdfjoin} no está instalado." && exit 1; }
-
-#Fechas y limites temporales
-dateDOWNLOAD=$(date -u +%Y%m%d%H%M)
-dateDOWNLOADu=$(date -u +%s000)
-weekDOWNLOAD=$(date '+%W')
-LIM=350 #maxímo número de segundos de descarga
-TIMEOUT=60 #máximo número de segundos intentando conectarse
+#TODO solo dejo el nombre del software sin el path (diferencia entre command y which?)
+software="Rscript pdfjoin"
+which ${software} > /dev/null 2>&1 || { echo "$(datePID): ${software} no está instalado." && exit 1; }
 
 #Directorios
 DIR_BASE=/home/cep/METAR
 DIR_DATA=$DIR_BASE/data/VALIDACION
 DIR_PLOTS=$DIR_DATA/PLOTS
+mkdir -p ${DIR_DATA} ${DIR_PLOTS}
 
 #Ficheros
 filegeonameID=${DIR_BASE}/geonameId-icao-correspondencia.txt #uso este fichero para que gonameID siempre correspoda al mismo METAR
 filemetarID=meteoflight_reports.php #se descarga
 
-#Comprobación de que existen los directorios y ficheros
-mkdir -p ${DIR_DATA} ${DIR_PLOTS}
+#scripts
+scriptR1=${DIR_BASE}/plot-PREDICvsMETAR.R
+scriptR2=${DIR_BASE}/mapa-errores-METAR.R
+
+#Comprobación de que existen los directorios, ficheros y scripts
+if [[ ! -e $scriptR1 ]]; then
+	echo "$(datePID): $scriptR1 no exite"
+  exit 1;
+fi
+if [[ ! -e $scriptR2 ]]; then
+	echo "$(datePID): $scriptR2 no exite"
+  exit 1;
+fi
 if [[ ! -e $filegeonameID ]]; then
 	echo "$(datePID): $filegeonameID no exite"
   exit 1;
 fi
-
-# Definición de fichero de bloqueo y de finalización.
-# Se comprueba si se está trabajando ya.
-#scriptName=$(basename $0 .sh)
-#lockFile=${DIR_DATA}/${scriptName}_${dateDOWNLOAD}.lock
-#if [[ -e ${lockFile} ]] ; then
-#  if [[ $(stat -c %Y ${lockFile}) -gt $(date -u +%s --date="180 minutes ago") ]] ; then
-#    exit # Problemas: el fichero ${lock} es muy antiguo.
-#  else
-#    touch ${lockFile}
-#  fi
-#fi
-
-###Descarga METARs info (LE españoles)
 if [[ -e ${filemetarID} ]] ; then
   rm -f $filemetarID
 fi
 
+#TODO ver si lockFile funciona
+# Definición de fichero de bloqueo y de finalización.
+# Se comprueba si se está trabajando ya.
+lockFile=${DIR_DATA}/${scriptName}_${dateDOWNLOAD}.lock
+if [[ -e ${lockFile} ]] ; then
+  if [[ $(stat -c %Y "${lockFile}") -gt $(date -u +%s --date="180 minutes ago") ]] ; then
+    echo "$(datePID): el fichero ${lockFile} es muy antiguo"
+    exit 1;
+  else
+    touch "${lockFile}"
+  fi
+fi
+
+#Fechas y limites temporales
+dateDOWNLOAD=$(date -u +%Y%m%d%H%M)
+#dateDOWNLOADu=$(date -u +%s000)
+weekDOWNLOAD=$(date '+%W')
+LIM=350 #maxímo número de segundos de descarga
+TIMEOUT=60 #máximo número de segundos intentando conectarse
+
+###Descarga METARs info: lon lat alt (LE españoles)
 echo "$(datePID): Inicia descarga de $filemetarID"
 curl -s -m $LIM --connect-timeout $TIMEOUT http://aire.ogimet.com/meteoflight_reports.php -o $filemetarID
 echo "$(datePID): Fin descarga de $filemetarID"
-metarID=$(cat $filemetarID | jq '.points[].icao' | grep 'LE[A-Z][A-Z]' | cut -c2-5)
-#Truco de Juan para asegurarme que tengo todos los json descargados adecaudamente
+
+metarID=$(cat "$filemetarID" | jq '.points[].icao' | grep 'LE[A-Z][A-Z]' | cut -c2-5)
+
+#Truco de Juan para asegurarme que tengo t-odos los json descargados adecaudamente
 ficherosdescargados=0
 while [ ${ficherosdescargados} -eq 0 ]
 do
@@ -76,26 +88,28 @@ do
     fnamePREDIC=${DIR_DATA}/PREDIC-$icao-${dateDOWNLOAD}
     fnameVALIDA=${DIR_DATA}/VALIDA-$icao-${weekDOWNLOAD}
 
-    if [ ! -e $fnamePREDIC.json ]
+    if [ ! -e "${fnamePREDIC}".json ]
     then
       echo "$(datePID): Ini descarga de json"
       ##Descarga METARs info (24h data) del json file
-      curl -s -m $LIM --connect-timeout $TIMEOUT http://aire.ogimet.com/meteoflight_reports.php?ICAO=$icao -o $fnameMETAR.json
+      curl -s -m $LIM --connect-timeout $TIMEOUT http://aire.ogimet.com/meteoflight_reports.php?ICAO=$icao -o "$fnameMETAR".json
       ##Descarga forecast info del json file en la localidad del METAR
+
       geonameId=$(cat $filegeonameID | grep "$icao" | awk '{print $2}')
       lat=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lat] | sort []')
       lon=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lon] | sort []')
       alt=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .alt] | sort []')
       curl -s -m $LIM --connect-timeout $TIMEOUT 'http://aguila.ogimet.com/cgi-bin/otf12?geonameId='$geonameId'&latitud='$lat'&longitud='$lon'&altitud='$alt'&zonaHoraria=etc/UTC&name='$icao'' -o $fnamePREDIC.json
       echo "$(datePID): Fin descarga de json"
+
       if [ $? -eq 0 ]
       then
-
+        #TODO no estoy usando fecha START ni fechaLEAD
 	      ## Uso la fecha inicial del fichero para hacerme una idea del LEAD time
-        fechaSTART=$(cat $fnamePREDIC.json | jq '.dias[0].utime.start')
-        fechaSTARTok=$(date --date=@$(echo $fechaSTART/1000 | bc) '+%Y%m%d%H%M')
+        #fechaSTART=$(cat $fnamePREDIC.json | jq '.dias[0].utime.start')
+        #fechaSTARTok=$(date --date=@$(echo $fechaSTART/1000 | bc) '+%Y%m%d%H%M')
 
-	      ##Para los METAR puede que haya pasos temporales sin datos de temp
+	      ##Para los METAR puede que haya pasos temporales sin datos de temp (con o sin NIL)
         fechasMETAR=$(jq '.metar_speci[].hour' ${fnameMETAR}.json)
         for idate in $fechasMETAR;do
           temp=$(jq '.metar_speci[] | select (.hour | contains('$idate')) | .report' ${fnameMETAR}.json | grep -o '[ M][0-9][0-9]/[M]*[0-9][0-9]' | tr "M" "-" | cut -c1-3)
@@ -103,47 +117,57 @@ do
           then
             fechaMETAR=$idate
             fechaMETARok=$(date --date=@$(echo $fechaMETAR/1000 | bc) '+%Y%m%d%H%M')
+            #Busco la predicción más cercana a la fecha del METAR
             fechaUP=$( echo $idate + 1800000 | bc)
             fechaDW=$( echo $idate - 1800000 | bc)
             fechaPRED=$(jq '.dias[].horas[] | select (.utime <= '$fechaUP' and .utime > '$fechaDW') | .utime' $fnamePREDIC.json)
             fechaDIF=$(echo '('$fechaPRED' - '$fechaMETAR') / 60000 ' | bc ) # minutos
-            fechaLEAD=$(echo '('$fechaMETAR' - '$fechaSTART') / 3600000' | bc ) #horas entre inicio de predic y METAR
+            #fechaLEAD=$(echo '('$fechaMETAR' - '$fechaSTART') / 3600000' | bc ) #horas entre inicio de predic y METAR
 
             tempPRED=$(jq '.dias[].horas[] | select (.utime == '$fechaPRED') | .temperatura.valor' $fnamePREDIC.json)
-            tempBIAS=$(echo $tempPRED - $temp | bc )
-           #echo $fechaDIF $fechaPRED $tempPRED $fechaMETAR $temp $tempBIAS $fechaMETARok $fechaSTARTok $fechaLEAD $icao
-           # echo $tempPRED $fechaUP $fechaDW $fechaPRED
+            #tempBIAS=$(echo $tempPRED - $temp | bc )
 
 	          #Nos aseguramos que los datos se comparan con la última pasada
             if [ -e $fnameVALIDA.txt ]
             then
               mv  $fnameVALIDA.txt kkrmdate
-              cat kkrmdate | awk 'NF==10' | grep -v $fechaMETAR  >  $fnameVALIDA.txt
+              cat kkrmdate | awk 'NF==6' | grep -v $fechaMETAR  >  $fnameVALIDA.txt
               rm kkrmdate
             fi
- 	          # Escribe datos en un archivo .txt para que pinte R (10 columnas!)
-            echo $fechaDIF $fechaPRED $tempPRED $fechaMETAR $temp $tempBIAS $fechaMETARok $fechaSTARTok $fechaLEAD $icao >> $fnameVALIDA.txt
+            # TODO simplificar, sobran columnas? NF
+ 	          # Escribe datos en un archivo .txt para que pinte R (5 columnas!)
+ 	          # R sólo va a usar $fechaPRED $tempPRED $fechaMETAR $temp
+            echo "$fechaDIF" "$fechaPRED" "$tempPRED" "$fechaMETAR" "$temp" "$fechaMETARok" >> "$fnameVALIDA".txt
           fi
 
         done
-
         echo "$(datePID): Ini plots w R $icao"
-        Rscript --vanilla ${DIR_BASE}/plot-PREDICvsMETAR.R $fnameVALIDA.txt ${DIR_PLOTS}/T-PREDIC-METAR-$icao-$weekDOWNLOAD.png $icao $fechaSTART ${DIR_DATA}/MEAN-BIAS-$icao-$weekDOWNLOAD.txt
+
+        #warning: si se cambian los parámetros de entradas modificar R consecuentemente
+        filein=$fnameVALIDA.txt
+        fileout=${DIR_DATA}/kkMEAN-BIAS-$icao-$weekDOWNLOAD.txt
+        plotout=${DIR_PLOTS}/T-PREDIC-METAR-$icao-$weekDOWNLOAD.png
+        Rscript --vanilla ${DIR_BASE}/plot-PREDICvsMETAR.R $filein $plotout $icao $lon $lat $fileout
 	      echo "$(datePID): Fin plots w R $icao"
 
-        cat $fnameVALIDA.txt | awk '{if (sqrt($6*$6) > 2 ) print "'$icao'",$0,'$fechaLEAD'}' >> ${DIR_DATA}/DIFGT2-predic-metar-$weekDOWNLOAD.txt
-        cat $fnameVALIDA.txt | awk '{print sqrt($6*$6), "'$icao'", "'$lon'", "'$lat'", $0,'$fechaLEAD'}' | sort -rn | head -1 >> ${DIR_DATA}/DIFMAX-predic-metar-$weekDOWNLOAD.txt
-
-      elif [ -e $fnamePRECIC.json ]
+      elif [ -e ${fnamePREDIC}.json ]
       then
         rm $fnamePREDIC.json
       fi
     fi
   done #end loop over icao
 
+  cat  ${DIR_DATA}/kkMEAN-BIAS-*-$weekDOWNLOAD.txt > ${DIR_DATA}/MAPA-T-PREDIC-METAR-$weekDOWNLOAD.txt
+  rm ${DIR_DATA}/kkMEAN-BIAS-*-$weekDOWNLOAD.txt
+
+  #warning: si se cambian los parámetros de entradas modificar R consecuentemente
+  filein=${DIR_DATA}/MAPA-T-PREDIC-METAR-$weekDOWNLOAD.txt
+  plotout=${DIR_PLOTS}/MAPA-T-PREDIC-METAR-$weekDOWNLOAD.pdf
+  Rscript --vanilla $scriptR2 "$filein"  $plotout
+
   ficherosdescargados=1
   for icao in $metarID; do
-    if [ ! -e $fnamePREDIC.json ]
+    if [ ! -e "$fnamePREDIC".json ]
     then
       ficherosdescargados=0;
       break;
@@ -151,11 +175,9 @@ do
   done
 done
 
-#rm -f ${lockFile}
-cat ${DIR_DATA}/MEAN-BIAS-*-$weekDOWNLOAD.txt > ${DIR_DATA}/JOINED-MEAN-BIAS-$weekDOWNLOAD.txt
-rm ${DIR_DATA}/MEAN-BIAS-*-$weekDOWNLOAD.txt
 #FIXME buscar como pegar los png
 #convert $DIR_DATA/PLOTS/*$weekDOWNLOAD.png $DIR_DATA/PLOTS/JOINED-T-PREDIC-METAR-$weekDOWNLOAD.png
-#FIXME cuando vea que funciona directamente borrar lo que hay en basura
+#FIXME cuando vea que funciona, directamente borrar lo que hay en basura
 mv $DIR_DATA/*.json ${DIR_DATA}/basura
+rm -f "${lockFile}"
 echo 'fin script predicciones'
